Implemented basic random walk
A simple implementation of a random walk, starting from a given node in
the graph has been added. As of right now, the flow of the final program
has not yet been added. The rungbims function acts as a "main" function,
which is used for testing purposes.
The random walk itself works fine, but could need some tweaks and
improvements in the future.